Example of how to group related instrumentations under a single module (AwsSdkModule)
