+++ dd-java-agent/instrumentation/aws-java-sdk-1.11.0/src/main/java/datadog/trace/instrumentation/aws/v0/RequestExecutorInstrumentation.java
@@ -6,9 +6,7 @@ import static datadog.trace.instrumentation.aws.v0.OnErrorDecorator.DECORATE;
 import static datadog.trace.instrumentation.aws.v0.OnErrorDecorator.SPAN_CONTEXT_KEY;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 
-import com.amazonaws.AmazonClientException;
 import com.amazonaws.Request;
-import com.amazonaws.handlers.RequestHandler2;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.bootstrap.instrumentation.api.AgentScope;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
@@ -16,29 +14,28 @@ import datadog.trace.bootstrap.instrumentation.api.AgentTracer;
 import net.bytebuddy.asm.Advice;
 
 /**
- * This is additional 'helper' to catch cases when HTTP request throws exception different from
- * {@link AmazonClientException} (for example an error thrown by another handler). In these cases
- * {@link RequestHandler2#afterError} is not called.
+ * Due to a change in the AmazonHttpClient class, this instrumentation is needed to support newer
+ * versions. The {@link AWSHttpClientInstrumentation} class should cover older versions.
  */
-public class AWSHttpClientInstrumentation
+public final class RequestExecutorInstrumentation
     implements Instrumenter.ForSingleType, Instrumenter.HasMethodAdvice {
 
   @Override
   public String instrumentedType() {
-    return "com.amazonaws.http.AmazonHttpClient";
+    return "com.amazonaws.http.AmazonHttpClient$RequestExecutor";
   }
 
   @Override
   public void methodAdvice(MethodTransformer transformer) {
     transformer.applyAdvice(
         isMethod().and(named("doExecute")),
-        AWSHttpClientInstrumentation.class.getName() + "$HttpClientAdvice");
+        RequestExecutorInstrumentation.class.getName() + "$RequestExecutorAdvice");
   }
 
-  public static class HttpClientAdvice {
+  public static class RequestExecutorAdvice {
     @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
     public static void methodExit(
-        @Advice.Argument(value = 0, optional = true) final Request<?> request,
+        @Advice.FieldValue("request") final Request<?> request,
         @Advice.Thrown final Throwable throwable) {
 
       final AgentScope scope = activeScope();
@@ -49,7 +46,7 @@ public class AWSHttpClientInstrumentation
         scope.close();
       }
 
-      if (throwable != null && request != null) {
+      if (throwable != null) {
         final AgentSpan span = request.getHandlerContext(SPAN_CONTEXT_KEY);
         if (span != null) {
           request.addHandlerContext(SPAN_CONTEXT_KEY, null);
